Add validation logic
This commit mostly adds logic to validate that dependencies are
accessible to the various components of the system.  This way
we will avoid applying breaking configuration changes to nodes
before their dependencies are available (e.g., don't configure
the master to use puppetdb until we've verified that the puppetdb
server is up and running).