Add conditional dependencies
Some of the classes in the module will conflict with one another
if they are applied on the same node but not in the correct order.
This commit is mostly about adding conditional dependencies
between these classes if they are deployed on the same node, so
that they will be executed in the proper order.